move TestReplaceSequence, TestRemoveIsolate to test_modify
