+++ tests/test_modify.py
@@ -1,4 +1,5 @@
 import subprocess
+from uuid import UUID
 from uuid import uuid4
 
 import pytest
@@ -6,6 +7,7 @@ from syrupy import SnapshotAssertion
 from syrupy.filters import props
 
 from ref_builder.repo import Repo
+from ref_builder.resources import RepoSequence
 from ref_builder.otu.create import create_otu
 from ref_builder.otu.update import (
     add_genbank_isolate,
@@ -13,8 +15,11 @@ from ref_builder.otu.update import (
     promote_otu_accessions,
     resize_monopartite_plan,
     set_isolate_plan,
+    replace_sequence_in_otu,
     set_representative_isolate,
+    delete_isolate_from_otu,
 )
+from ref_builder.utils import IsolateName, IsolateNameType
 from ref_builder.plan import (
     MonopartitePlan,
     MultipartitePlan,
@@ -25,6 +30,7 @@ from ref_builder.plan import (
 
 
 def test_update_representative_isolate(scratch_repo: Repo):
+    """Test representative isolate replacement."""
     taxid = 345184
 
     otu_before = scratch_repo.get_otu_by_taxid(taxid)
@@ -234,6 +240,68 @@ class TestUpdateRepresentativeIsolateCommand:
 
         assert otu_after.repr_isolate == otu_before.get_isolate_id_by_name(
             repr_isolate_name_after
+        )
+
+
+class TestRemoveIsolate:
+    def test_ok(self, scratch_repo):
+        """Test that a given isolate can be removed from the OTU."""
+        taxid = 1169032
+
+        otu_before = scratch_repo.get_otu_by_taxid(taxid)
+
+        isolate_id = otu_before.get_isolate_id_by_name(
+            IsolateName(type=IsolateNameType.ISOLATE, value="WMoV-6.3"),
+        )
+
+        assert type(isolate_id) is UUID
+
+        delete_isolate_from_otu(scratch_repo, otu_before, isolate_id)
+
+        otu_after = scratch_repo.get_otu_by_taxid(taxid)
+
+        assert isolate_id not in otu_after.isolate_ids
+
+        assert otu_after.get_isolate(isolate_id) is None
+
+        assert otu_before.get_isolate(isolate_id).accessions not in otu_after.accessions
+
+        assert len(otu_after.isolate_ids) == len(otu_before.isolate_ids) - 1
+
+
+class TestReplaceSequence:
+    def test_ok(self, precached_repo):
+        """Test sequence replacement and deletion."""
+        otu_before = create_otu(
+            precached_repo,
+            1169032,
+            ["MK431779"],
+            acronym="",
+        )
+
+        isolate_id, old_sequence_id = (
+            otu_before.get_sequence_id_hierarchy_from_accession(
+                "MK431779",
+            )
+        )
+
+        assert type(old_sequence_id) is UUID
+
+        sequence = replace_sequence_in_otu(
+            repo=precached_repo,
+            otu=otu_before,
+            new_accession="NC_003355",
+            replaced_accession="MK431779",
+        )
+
+        assert type(sequence) is RepoSequence
+
+        otu_after = precached_repo.get_otu_by_taxid(1169032)
+
+        assert (
+            otu_after.accessions
+            == otu_after.get_isolate(isolate_id).accessions
+            == {"NC_003355"}
         )
